requirements.txt
# Core dependencies
gitpython==3.1.41
anthropic>=0.9.0
python-dotenv==1.0.1

# Git operations
GitPython==3.1.41

# Additional utilities
pathlib==1.0.1
typing-extensions>=4.5.0

# Python version requirement
# Requires Python >= 3.8